Provide hugepage advice only when the THP enabled setting is madvise
When THP is enabled system-wide, there is no need to provide hugepage
advice to enable THP for a region of memory.  On the other hand, when
the THP defragmentation mode is defer+madvise, hugepage advice can
result in unwanted application stalls to compact and defragment memory
if THP is not immediately available for allocation.